+++ android/app/build.gradle
@@ -32,7 +32,7 @@ android {
         targetCompatibility JavaVersion.VERSION_17
     }
     kotlinOptions {
-        jvmTarget = '1.8'
+        jvmTarget = '17'
     }
     dataBinding {
         enabled = true
