+++ android/app/src/main/AndroidManifest.xml
@@ -13,18 +13,13 @@
         android:theme="@style/Theme.DigginRoom"
         tools:targetApi="31">
         <activity
-            android:name=".RoomActivity"
+            android:name=".MainActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
-        </activity>
-        <activity
-            android:name=".MainActivity"
-            android:exported="true">
-
         </activity>
     </application>
