+++ android/app/src/main/res/layout/item_recycler_view.xml
